Implement a filename exposing raw cache mode
This intends to support zerocopy/SendFile scenarios. There
is no serialization in this mode, the cache is basically
managing a bunch of files.

The mode intentionally only accepts byte[] or streams,
it does not support fetching them to avoid memory and
stream lifetime issues.

Because of the lifetime issue, a configurable
safety margin is added to the expiry logic.